refactor(mobile): Change package architecture for mobile app
