+++ mobile/app/src/main/res/menu/bottom_nav_driver.xml
@@ -2,14 +2,20 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/nav_dashboard"
-        android:title="Dashboard"
-        android:icon="@android:drawable/ic_menu_view" />
+        android:id="@+id/nav_current_driver"
+        android:title="Current"
+        android:icon="@drawable/ic_current_ride" />
 
     <item
         android:id="@+id/nav_history_driver"
         android:title="History"
         android:icon="@drawable/ic_history" />
+
+    <item
+        android:id="@+id/nav_dashboard"
+        android:title="Dashboard"
+        android:icon="@android:drawable/ic_menu_view" />
+
 
     <item
         android:id="@+id/nav_chat"
